Schema_LOADER/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.34.3
PyYAML==6.0.1
pydantic==2.5.3 